third_party/WebKit/LayoutTests/geolocation-api/permission-denied-already-clear-watch.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/geolocation-mock.js"></script>
</head>
<body>
<script>
description("Tests that when Geolocation permission has been denied prior to a call to watchPosition, and the watch is cleared in the error callback, there is no crash. This a regression test for https://bugs.webkit.org/show_bug.cgi?id=32111.");

var error;

geolocationMock.then(mock => {

    // Prime the Geolocation instance by denying permission.
    mock.setGeolocationPermission(false);
    mock.setGeolocationPosition(51.478, -0.166, 100);

    navigator.geolocation.getCurrentPosition(function(p) {
        testFailed('Success callback invoked unexpectedly');
        finishJSTest();
    }, function(e) {
        error = e;
        shouldBe('error.code', 'error.PERMISSION_DENIED');
        shouldBe('error.message', '"User denied Geolocation"');
        debug('');
        continueTest();
    });

    function continueTest()
    {
        // Make another request, with permission already denied.
        var watchId = navigator.geolocation.watchPosition(function(p) {
            testFailed('Success callback invoked unexpectedly');
            finishJSTest();
        }, function(e) {
            error = e;
            shouldBe('error.code', 'error.PERMISSION_DENIED');
            shouldBe('error.message', '"User denied Geolocation"');
            navigator.geolocation.clearWatch(watchId);
            window.setTimeout(finishJSTest, 0);
        });
    }
});

window.jsTestIsAsync = true;
</script>
</body>
</html>
